+++ src/api/logic_traits/access_logic_trait.rs
@@ -1,0 +1,26 @@
+use crate::api::server::server::{
+    CreateAccessRequest, DeleteAccessRequest, ListAccessInfoRequest, ListAccessInfoResponse,
+    UpdateAccessRequest,
+};
+use async_trait::async_trait;
+use tonic::{Request, Response, Status};
+
+#[async_trait]
+pub trait AccessLogicTrait {
+    async fn list_access_info(
+        &self,
+        request: Request<ListAccessInfoRequest>,
+    ) -> Result<Response<ListAccessInfoResponse>, Status>;
+    async fn create_access(
+        &self,
+        request: Request<CreateAccessRequest>,
+    ) -> Result<Response<()>, Status>;
+    async fn update_access(
+        &self,
+        request: Request<UpdateAccessRequest>,
+    ) -> Result<Response<()>, Status>;
+    async fn delete_access(
+        &self,
+        request: Request<DeleteAccessRequest>,
+    ) -> Result<Response<()>, Status>;
+}
