+++ src/api/logic_traits/session_logic_trait.rs
@@ -1,0 +1,12 @@
+use crate::api::server::server::{GetAuthTokenRequest, GetAuthTokenResponse};
+use async_trait::async_trait;
+use tonic::{Request, Response, Status};
+
+#[async_trait]
+pub trait SessionLogicTrait {
+    async fn delete_session(&self, _request: Request<()>) -> Result<Response<()>, Status>;
+    async fn get_auth_token(
+        &self,
+        request: Request<GetAuthTokenRequest>,
+    ) -> Result<Response<GetAuthTokenResponse>, Status>;
+}
