+++ src/api/mod.rs
@@ -1,6 +1,7 @@
 pub mod auth;
 pub mod context_collection;
 pub mod ecdar_api;
-pub mod hashing_context;
-pub mod reveaal_context;
+pub mod logic_collection;
+pub mod logic_impls;
+pub mod logic_traits;
 pub mod server;
