+++ src/database/context_impls/hashing_context.rs
@@ -1,9 +1,5 @@
+use crate::api::logic_traits::HashingContextTrait;
 use bcrypt::{hash, verify, DEFAULT_COST};
-
-pub trait HashingContextTrait: Send + Sync {
-    fn hash_password(&self, password: String) -> String;
-    fn verify_password(&self, password: String, hash: &str) -> bool;
-}
 
 pub struct HashingContext;
