+++ src/database/context_traits/hashing_context_trait.rs
@@ -1,0 +1,4 @@
+pub trait HashingContextTrait: Send + Sync {
+    fn hash_password(&self, password: String) -> String;
+    fn verify_password(&self, password: String, hash: &str) -> bool;
+}
