+++ src/database/context_traits/mod.rs
@@ -1,6 +1,7 @@
 pub mod access_context_trait;
 pub mod database_context_trait;
 pub mod entity_context_trait;
+pub mod hashing_context_trait;
 pub mod in_use_context_trait;
 pub mod project_context_trait;
 pub mod query_context_trait;
@@ -10,6 +11,7 @@ pub mod user_context_trait;
 pub use access_context_trait::AccessContextTrait;
 pub use database_context_trait::DatabaseContextTrait;
 pub use entity_context_trait::EntityContextTrait;
+pub use hashing_context_trait::HashingContextTrait;
 pub use in_use_context_trait::InUseContextTrait;
 pub use project_context_trait::ProjectContextTrait;
 pub use query_context_trait::QueryContextTrait;
