+++ src/tests/api/user_logic.rs
@@ -1,8 +1,8 @@
-use crate::api::server::server::ecdar_api_auth_server::EcdarApiAuth;
-use crate::api::server::server::ecdar_api_server::EcdarApi;
+use crate::api::logic_impls::{QueryLogic, UserLogic};
+use crate::api::logic_traits::UserLogicTrait;
 use crate::api::server::server::{CreateUserRequest, GetUsersRequest, UpdateUserRequest};
 use crate::entities::user;
-use crate::tests::api::helpers::{get_mock_concrete_ecdar_api, get_mock_services};
+use crate::tests::api::helpers::{disguise_mocks, get_mock_contexts};
 use mockall::predicate;
 use sea_orm::DbErr;
 use std::str::FromStr;
@@ -10,15 +10,16 @@ use tonic::{metadata, Code, Request};
 
 #[tokio::test]
 async fn delete_user_nonexistent_user_returns_err() {
-    let mut mock_services = get_mock_services();
+    let mut mock_contexts = get_mock_contexts();
 
-    mock_services
+    mock_contexts
         .user_context_mock
         .expect_delete()
         .with(predicate::eq(1))
         .returning(|_| Err(DbErr::RecordNotFound("".into())));
 
-    let api = get_mock_concrete_ecdar_api(mock_services);
+    let contexts = disguise_mocks(mock_contexts);
+    let user_logic = UserLogic::new(contexts);
 
     let mut delete_request = Request::new(());
 
@@ -27,7 +28,7 @@ async fn delete_user_nonexistent_user_returns_err() {
         .metadata_mut()
         .insert("uid", metadata::MetadataValue::from_str("1").unwrap());
 
-    let delete_response = api.delete_user(delete_request).await.unwrap_err();
+    let delete_response = user_logic.delete_user(delete_request).await.unwrap_err();
     let expected_response_code = Code::Internal;
 
     assert_eq!(delete_response.code(), expected_response_code);
@@ -35,7 +36,7 @@ async fn delete_user_nonexistent_user_returns_err() {
 
 #[tokio::test]
 async fn delete_user_existing_user_returns_ok() {
-    let mut mock_services = get_mock_services();
+    let mut mock_contexts = get_mock_contexts();
 
     let user = user::Model {
         id: 1,
@@ -44,13 +45,14 @@ async fn delete_user_existing_user_returns_ok() {
         password: "".to_string(),
     };
 
-    mock_services
+    mock_contexts
         .user_context_mock
         .expect_delete()
         .with(predicate::eq(1))
         .returning(move |_| Ok(user.clone()));
 
-    let api = get_mock_concrete_ecdar_api(mock_services);
+    let contexts = disguise_mocks(mock_contexts);
+    let user_logic = UserLogic::new(contexts);
 
     let mut delete_request = Request::new(());
 
@@ -59,14 +61,14 @@ async fn delete_user_existing_user_returns_ok() {
         .metadata_mut()
         .insert("uid", metadata::MetadataValue::from_str("1").unwrap());
 
-    let delete_response = api.delete_user(delete_request).await;
+    let delete_response = user_logic.delete_user(delete_request).await;
 
     assert!(delete_response.is_ok());
 }
 
 #[tokio::test]
 async fn create_user_nonexistent_user_returns_ok() {
-    let mut mock_services = get_mock_services();
+    let mut mock_contexts = get_mock_contexts();
 
     let password = "Password123".to_string();
 
@@ -83,26 +85,27 @@ async fn create_user_nonexistent_user_returns_ok() {
         password: password.clone(),
     });
 
-    mock_services
+    mock_contexts
         .hashing_context_mock
         .expect_hash_password()
         .returning(move |_| password.clone());
 
-    mock_services
+    mock_contexts
         .user_context_mock
         .expect_create()
         .with(predicate::eq(user.clone()))
         .returning(move |_| Ok(user.clone()));
 
-    let api = get_mock_concrete_ecdar_api(mock_services);
+    let contexts = disguise_mocks(mock_contexts);
+    let user_logic = UserLogic::new(contexts);
 
-    let create_user_response = api.create_user(create_user_request).await;
+    let create_user_response = user_logic.create_user(create_user_request).await;
     assert!(create_user_response.is_ok());
 }
 
 #[tokio::test]
 async fn create_user_duplicate_email_returns_error() {
-    let mut mock_services = get_mock_services();
+    let mut mock_contexts = get_mock_contexts();
 
     let password = "Password123".to_string();
 
@@ -119,28 +122,30 @@ async fn create_user_duplicate_email_returns_error() {
         password: password.clone(),
     });
 
-    mock_services
+    mock_contexts
         .hashing_context_mock
         .expect_hash_password()
         .returning(move |_| password.clone());
 
-    mock_services
+    mock_contexts
         .user_context_mock
         .expect_create()
         .with(predicate::eq(user.clone()))
         .returning(move |_| Err(DbErr::RecordNotInserted)); //todo!("Needs to be a SqlError with UniqueConstraintViolation with 'email' in message)
 
-    let api = get_mock_concrete_ecdar_api(mock_services);
+    let contexts = disguise_mocks(mock_contexts);
+    let user_logic = UserLogic::new(contexts);
 
-    let res = api.create_user(create_user_request).await;
+    let res = user_logic.create_user(create_user_request).await;
     assert_eq!(res.unwrap_err().code(), Code::Internal); //todo!("Needs to be code AlreadyExists when mocked Error is corrected)
 }
 
 #[tokio::test]
 async fn create_user_invalid_email_returns_error() {
-    let mock_services = get_mock_services();
+    let mock_contexts = get_mock_contexts();
 
-    let api = get_mock_concrete_ecdar_api(mock_services);
+    let contexts = disguise_mocks(mock_contexts);
+    let user_logic = UserLogic::new(contexts);
 
     let create_user_request = Request::new(CreateUserRequest {
         email: "invalid-email".to_string(),
@@ -148,13 +153,13 @@ async fn create_user_invalid_email_returns_error() {
         password: "123".to_string(),
     });
 
-    let res = api.create_user(create_user_request).await;
+    let res = user_logic.create_user(create_user_request).await;
     assert_eq!(res.unwrap_err().code(), Code::InvalidArgument);
 }
 
 #[tokio::test]
 async fn create_user_duplicate_username_returns_error() {
-    let mut mock_services = get_mock_services();
+    let mut mock_contexts = get_mock_contexts();
 
     let password = "Password123".to_string();
 
@@ -171,28 +176,30 @@ async fn create_user_duplicate_username_returns_error() {
         password: password.clone(),
     });
 
-    mock_services
+    mock_contexts
         .hashing_context_mock
         .expect_hash_password()
         .returning(move |_| password.clone());
 
-    mock_services
+    mock_contexts
         .user_context_mock
         .expect_create()
         .with(predicate::eq(user.clone()))
         .returning(move |_| Err(DbErr::RecordNotInserted)); //todo!("Needs to be a SqlError with UniqueConstraintViolation with 'username' in message)
 
-    let api = get_mock_concrete_ecdar_api(mock_services);
+    let contexts = disguise_mocks(mock_contexts);
+    let user_logic = UserLogic::new(contexts);
 
-    let res = api.create_user(create_user_request).await;
+    let res = user_logic.create_user(create_user_request).await;
     assert_eq!(res.unwrap_err().code(), Code::Internal); //todo!("Needs to be code AlreadyExists when mocked Error is corrected)
 }
 
 #[tokio::test]
 async fn create_user_invalid_username_returns_error() {
-    let mock_services = get_mock_services();
+    let mock_contexts = get_mock_contexts();
 
-    let api = get_mock_concrete_ecdar_api(mock_services);
+    let contexts = disguise_mocks(mock_contexts);
+    let user_logic = UserLogic::new(contexts);
 
     let create_user_request = Request::new(CreateUserRequest {
         email: "valid@email.com".to_string(),
@@ -200,13 +207,13 @@ async fn create_user_invalid_username_returns_error() {
         password: "123".to_string(),
     });
 
-    let res = api.create_user(create_user_request).await;
+    let res = user_logic.create_user(create_user_request).await;
     assert_eq!(res.unwrap_err().code(), Code::InvalidArgument);
 }
 
 #[tokio::test]
 async fn create_user_valid_request_returns_ok() {
-    let mut mock_services = get_mock_services();
+    let mut mock_contexts = get_mock_contexts();
 
     let password = "Password123".to_string();
 
@@ -223,26 +230,27 @@ async fn create_user_valid_request_returns_ok() {
         password: password.clone(),
     });
 
-    mock_services
+    mock_contexts
         .hashing_context_mock
         .expect_hash_password()
         .returning(move |_| password.clone());
 
-    mock_services
+    mock_contexts
         .user_context_mock
         .expect_create()
         .with(predicate::eq(user.clone()))
         .returning(move |_| Ok(user.clone()));
 
-    let api = get_mock_concrete_ecdar_api(mock_services);
+    let contexts = disguise_mocks(mock_contexts);
+    let user_logic = UserLogic::new(contexts);
 
-    let create_user_response = api.create_user(create_user_request).await;
+    let create_user_response = user_logic.create_user(create_user_request).await;
     assert!(create_user_response.is_ok());
 }
 
 #[tokio::test]
 async fn update_user_returns_ok() {
-    let mut mock_services = get_mock_services();
+    let mut mock_contexts = get_mock_contexts();
 
     let old_user = user::Model {
         id: 1,
@@ -258,25 +266,26 @@ async fn update_user_returns_ok() {
         password: "g76df2gd7hd837g8hjd8723hd8gd823d82d3".to_string(),
     };
 
-    mock_services
+    mock_contexts
         .user_context_mock
         .expect_get_by_id()
         .with(predicate::eq(1))
         .returning(move |_| Ok(Some(old_user.clone())));
 
-    mock_services
+    mock_contexts
         .hashing_context_mock
         .expect_hash_password()
         .with(predicate::eq("StrongPassword123".to_string()))
         .returning(move |_| "g76df2gd7hd837g8hjd8723hd8gd823d82d3".to_string());
 
-    mock_services
+    mock_contexts
         .user_context_mock
         .expect_update()
         .with(predicate::eq(new_user.clone()))
         .returning(move |_| Ok(new_user.clone()));
 
-    let api = get_mock_concrete_ecdar_api(mock_services);
+    let contexts = disguise_mocks(mock_contexts);
+    let user_logic = UserLogic::new(contexts);
 
     let mut update_user_request = Request::new(UpdateUserRequest {
         email: Some("newuser@example.com".to_string()),
@@ -288,22 +297,23 @@ async fn update_user_returns_ok() {
         .metadata_mut()
         .insert("uid", metadata::MetadataValue::from_str("1").unwrap());
 
-    let update_user_response = api.update_user(update_user_request).await;
+    let update_user_response = user_logic.update_user(update_user_request).await;
 
     assert!(update_user_response.is_ok())
 }
 
 #[tokio::test]
 async fn update_user_non_existant_user_returns_err() {
-    let mut mock_services = get_mock_services();
+    let mut mock_contexts = get_mock_contexts();
 
-    mock_services
+    mock_contexts
         .user_context_mock
         .expect_get_by_id()
         .with(predicate::eq(1))
         .returning(move |_| Err(DbErr::RecordNotFound("".to_string())));
 
-    let api = get_mock_concrete_ecdar_api(mock_services);
+    let contexts = disguise_mocks(mock_contexts);
+    let user_logic = UserLogic::new(contexts);
 
     let mut update_user_request = Request::new(UpdateUserRequest {
         email: Some("new_test@test".to_string()),
@@ -315,14 +325,14 @@ async fn update_user_non_existant_user_returns_err() {
         .metadata_mut()
         .insert("uid", metadata::MetadataValue::from_str("1").unwrap());
 
-    let res = api.update_user(update_user_request).await;
+    let res = user_logic.update_user(update_user_request).await;
 
     assert_eq!(res.unwrap_err().code(), Code::Internal);
 }
 
 #[tokio::test]
 async fn get_users_returns_ok() {
-    let mut mock_services = get_mock_services();
+    let mut mock_contexts = get_mock_contexts();
 
     let users = vec![
         user::Model {
@@ -339,36 +349,38 @@ async fn get_users_returns_ok() {
         },
     ];
 
-    mock_services
+    mock_contexts
         .user_context_mock
         .expect_get_by_ids()
         .returning(move |_| Ok(users.clone()));
 
-    let api = get_mock_concrete_ecdar_api(mock_services);
+    let contexts = disguise_mocks(mock_contexts);
+    let user_logic = UserLogic::new(contexts);
 
     let get_users_request = Request::new(GetUsersRequest { ids: vec![1, 2] });
 
-    let get_users_response = api.get_users(get_users_request).await.unwrap();
+    let get_users_response = user_logic.get_users(get_users_request).await.unwrap();
 
     assert_eq!(get_users_response.get_ref().users.len(), 2);
 }
 
 #[tokio::test]
 async fn get_users_returns_empty_array() {
-    let mut mock_services = get_mock_services();
+    let mut mock_contexts = get_mock_contexts();
 
     let users: Vec<user::Model> = vec![];
 
-    mock_services
+    mock_contexts
         .user_context_mock
         .expect_get_by_ids()
         .returning(move |_| Ok(users.clone()));
 
-    let api = get_mock_concrete_ecdar_api(mock_services);
+    let contexts = disguise_mocks(mock_contexts);
+    let user_logic = UserLogic::new(contexts);
 
     let get_users_request = Request::new(GetUsersRequest { ids: vec![1, 2] });
 
-    let get_users_response = api.get_users(get_users_request).await.unwrap();
+    let get_users_response = user_logic.get_users(get_users_request).await.unwrap();
 
     assert_eq!(get_users_response.get_ref().users.len(), 0);
 }
